+++ workspaces/liferay-partner-workspace/client-extensions/liferay-partner-etc-spring-boot/src/main/resources/application-default.properties
@@ -12,7 +12,7 @@ com.liferay.lxc.dxp.server.protocol=http
 
 liferay-partner-etc-spring-boot-oauth-application-headless-server.oauth2.headless.server.client.secret=myfancypassword
 liferay.oauth.application.external.reference.codes=liferay-partner-etc-spring-boot-oauth-application-headless-server
-liferay.oauth.urls.excludes=/ready, /statusManagement
+liferay.oauth.urls.excludes=/object/action/mdf-claim/status-management,/ready
 
 #
 # Spring Boot
